Bugfix: IKI variable substitution results in incorrect parameter structure for reserved parameters.
The commit 9ebfcc071f11f82b435cfa0f73e4961912571c38 didn't address this case.

There is an existing "separate = F_true;" that suggest that I was planning on working on this.
I probably overlooked and forgot about this block.

Do not increment the parameters array and instead prepend a single space only when needed.