Feature: The fake program should support the "current" parameter.
This is the compliment to the "top" variable parameter.

The current working directory, also known as the present working directory, should be an IKI variable.

Use the f_string_dynamic_append_assure() rather than f_string_dynamic_append() to only add the trailing slash if it is not already present.

Update the documentation.