Bugfix: IKI substitution is incorrectly applied.
The function controller_rule_action_read_delimit_apply() is remapping the array and applies the delimits.
Another function is then applying the same delimits again, but against the remapped buffer.
This results in the delimits being applied to the wrong locations in the string.

Get rid of the controller_rule_action_read_delimit_apply().
The use of this function results in an incorrect string when printing validation.

Thoroughly clear the entire IKI data when processing the actions.