Bugfix: Copy should destination should be the base name rather than the entire path.
The entire path is being appended to the destination when performing a clone or a copy operation.
Example:
  copy a/b/c.txt destination/

This should copy c.txt as "destination/c.txt".
What instead happens is this: "destination/a/b/c.txt".

Change the code to now detect the file base name and append the base name rather than append the entire path.