Bugfix: Link operation is incorrect, support new link arguments, and clone and copy operation failure problems.
The link operation target path is relative to the point path.
The previous code is trying to treat the target path in isolation.
Change the behavior to make the target path relative to the point path.

Add two new options to make creating symbolic linking easier to use:
1) "force": Used to forcibly overwrite an existing file or directory.
2) "strict": Used to require the target path to exist when creating the symbolic links.

The clone and copy operations now have better error return code processing.