Update: Add +E/++error parameter to firewall.
The previous security commit 7348813d0af17acda440eb66c038f6bdd07c1a6e is the result of an incomplete addition of the +E/++error parameter.
The size of the total parameters was incremented to ensure space for the +E/++error but the +E/++error was never added.

This increments the total parameters and adds the +E/++error row.