Update: Add missing structures to f_account_t and implement allocation functions.
I failed to notice that f_account is incomplete and this failed to make the 0.6 release window.
This is an API change and will only be applied to the 0.7 development releases as a result.

Provide real functions rather than macros for allocation and deallocation of the f_account_t structure.
Make sure both f_accounts_t and f_acountss_t exist for standard consistency.
Add unit tests.