Feature: Add support for 'stand_alone' builds in the package script.
Provide a way to compile programs, such as Featureless Make, with all of their dependencies and most (if not all) of the unnecessary dependencies not part of the programs.
These programs, such as the Featureless Make program, are compiled as a single program with no libfll.so (or program library such as libfake.so) built.

These are still compiled as either shared or static.
All non-FLL libraries in the shared programs are still linked as libraries.

This is a feature in the helper script and does not result in any API (or ABI) breakage.