Update: Add support for stand_alone builds in the example bootstrap script.
The bootstrap-example.sh script now supports building in "stand alone" mode for any program that has the mode defined.
The currently supported programs "fake" and "utf8" are still hard coded.
Future versions may check for the existence of the named files rather than being hard coded.

Update utf8 to compile.
Make the same changes for fake, but fake is not yet read because it has not been migrated yet.