Bugfix: The iki_read is should apply the "--wrap" when using "--substitute".
Given some iki text such as:
  This is some random:"iki" random:"text".

Run the command:
  # echo 'This is some random:"iki" random:"text".' | iki_read -W random "___" "______" -L -s random text PIZZA -r random tomato

The result that I get is:
  random:"___tomato______"
  random:"PIZZA"

I think the following is better:
  random:"___tomato______"
  random:"___PIZZA______"

This makes more sense to me rather than having substitute not include wrap.
This is a behavioral change that I thought and perhaps still think can be considered a bug.
This "bug" is by design as per the documentation but in retrospect I think this is a mistake in the design and should be considered a bug.