Feature: The iki_read should still support the older "--substitute" behavior.
The commit bbc7591a040baeca7d19d1d6b644b80ad2e70d01 considered the behavior of --substitute to be incorrect and therefore a bug.

After some consideration I decided that this "bug" would still be useful in certain cases.
The old behavior of "--substitute" where the "--wrap" is ignored is re-created within the new parameter "--reassign".

This is a grey area when it comes to a "bug" and a "feature".
The behavior that is considered a "bug" turned out to be useful and as the behavior is already present it is carried on.