Update: Support Unicode 15.0.
I couldn't easily find what Unicode version I am trying to suppot so I added a unicode.txt document.

The initial work was done for Unicode 12.1 and later switched to 14.0.
I was still in the initial stages of learning Unicode and so I expect there to be notable incompleteness or incorrectness.
With the release of Unicode 15.0 I am able to determine a distinct set of changes and I have applied the changes.

Add a private_inline_f_utf_character_handle_digit_offset() inline function to simplify some of the redundant code.