Update: Slight improvement in UTF unassigned detection and add or correct some ranges.
Change "character" to "sequence" to be consistent with the rest of the private functions.

Add several additional checks to ideally reduce the number of checks.
This is only a cursory pass and not much effort is put into fine-tuned optimization.
There is huge room for improvement.

I discovered some ranges that needed to be added as unassined and added them.
Include the invalid ranges as unassigned.