Update: Finish program related work for fss_payload_write.
This appears to be done now.
The next step for fss_payload_write is to use it as the starting point for the fss_write program that is intended to provide all of the fss_*_write programs.

I will likely make improvements to the lazy approach I took regarding loading the entire file when processing.

I found two bugs regarding the pipe processing.
1) The first character never gets printed for payload Content.
2) An end-of-content character must be printed after the Content if printing partial Objects or if the "payload" Content is not yet printed.