Progress: Stub out the fss_write program.
The fss_write program is going to replace all of the other fss_X_write programs.
This is the initial setup and may not be the final structure.
I still have some fiddlign to do to figure out how I want them to share code.

The final result is intended to have the fss_write being capable of compiling all of the fss_X_write programs as well as the fss_write program.