Bugfix: Invalid print of character when a placeholder should be printed in byte_dump.
The following is happening:
  # clear ; echo -n "xa" | byte_dump -wt 2 && echo -n "∩" | byte_dump -wt 2 && echo -n "∩xa" | byte_dump -wt 2

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  78 61  | x a  |

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  e2 88  | ∩    |
  0000000000000001  a9     | ∩    |

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  e2 88  | ∩    |
  0000000000000001  a9 78  |   x  |
  0000000000000002  61     | a    |

In the second case the line 0000000000000001 should not print the string '| ∩    |'.
This is happening because the character is not being properly reset in the situation where the overflow happens at the end of input.

With this change the results should now be:
  # clear ; echo -n "xa" | byte_dump -wt 2 && echo -n "∩" | byte_dump -wt 2 && echo -n "∩xa" | byte_dump -wt 2

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  78 61  | x a  |

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  e2 88  | ∩    |
  0000000000000001  a9     |      |

  Piped Byte Dump: (in Hexidecimal)
  0000000000000000  e2 88  | ∩    |
  0000000000000001  a9 78  |   x  |
  0000000000000002  61     | a    |

Perform a trivial cleanup/optimization where inverse of bytes is being checked for.
This is pointless when there is an else block.
Reverse the order and remove the "not" operator.