Update: Add support for backtick quoting for IKI.
Comply with the recent addition to the IKI specification that designates backticks as a quote alternative to single and double quotes.

Fix several inconsistencies observed, such as "quoted" vs "quote".

The code is inconsistently using a type code and a literal character for passing the desired quote.
This is probably a long lost incomplete refactor.
Change all of the FSS code to accept the character as a uint8_t.
Get rid of the f_fss_quote_t and related, replacing them with uint8_t.

Get rid of the f_fss_delimit_* macros.
These are redundant and it is simpler just to use the f_fss_* variants.

Rename F_fss_delimit_placeholder_s to F_string_placeholder_s (and similar).

The flags for single and double quotes are not needed when the charaacter is being explicitly set.