Update: More consistently follow the recent clarifications of the style guide for defines.
I discovered that the "_s" type of notation to be incredibly helpful while working on the stand alone projects.
I also noticed some inconsistency in the naming structure to some extent.

I updated and clarified the style guide and then updated the defines where needed.

This turned out to be more work than expect and has been rushed a little.
Look out for mistakes.