Update: Redesign project structure to follow the "main" design used in fss_write.
This makes the program structure more consistent.
The structure being followed easily allows for a more object-oriented directory structure and design.
This design is more accurately closer to "progam-oriented" than "object-oriented" due to the structure revolving around a program.

The design of the code itself in fss_write is object-oriented like.
Use this code design for fss_status_code and status_code.
The other programs do not need to do this and have everything under the "main".