Feature: Add support for printing partial data with or without closing data in fss_write.
The "complete" "partial" FSS write enumaration settings still prints the end of Object or Content data.
This "partial" simply does not print the end of line characters.

Change the behavior to act like the newly added feature regarting "complete" "trim".
Provide parameters to toggle on printing of the closing data.