Update: Slight changes to design around the build settings arguments.
Improve the argument handling when the fakefile contains a build settings with arguments.

This is done in observance of a bug where passed modes are combined into a single parameter when modes are passed via an IKI replacement.
A follow up bugfix that may involve some refactoring or functionality reinterpretation is needed.

This uses an explicit array when needed rather than an inexplicit pointer.

This repeats the parsing when printing the output so that information output is not misleading.