Cleanup: Bring a copy of time.txt from Controller to the core project.
New functionality needs to be added to the package.sh script to conditionally add documentation and specifications to any given package.
Once this is done, then the now redundant time.txt can be removed from the Controller and possibly Control projects.

There also needs to be a formally specification file for time.txt and not just documentation.