Progress: Prototype the callback paameter for console parameters.
There needs to be a way to have conditional parameters and other relationships.
Provide an optional callback to help facilitate this without adding any additional complicated structures (other than the callback itself).

I need to figure out how I want the arguments to the callback to be defined.
This stubs out the design.
The parameters to the callback are subject to change once I determine what they should be.